globalorderid等应该使用int32类型目前int8， 这样会限制理论的进程数量 https://github.com/kbengine/kbengine/issues/178

+++ kbe/src/lib/common/platform.hpp
@@ -321,7 +321,7 @@ typedef int32													ENTITY_ID;												// entityID的类型
 typedef uint32													SPACE_ID;												// 一个space的id
 typedef uint32													CALLBACK_ID;											// 一个callback由CallbackMgr分配的id
 typedef uint64													COMPONENT_ID;											// 一个服务器组件的id
-typedef int8													COMPONENT_ORDER;										// 一个组件的启动顺序
+typedef int32													COMPONENT_ORDER;										// 一个组件的启动顺序
 typedef	uint32													TIMER_ID;												// 一个timer的id类型
 typedef uint8													MAIL_TYPE;												// mailbox 所投递的mail类别的类别
 typedef uint32													GAME_TIME;
